+++ src/js/components/InputImage.vue
@@ -1,31 +1,45 @@
 <template>
   <div class="file-image-wrap">
-    <figure class="image" v-if="value">
-      <img :src="value" :title="value">
-      <button class="delete" v-on:click="removeFile()" :disabled="disabled || isProcessing">remove</button>
+    <figure class="image" v-if="!!value || !!imageBase64">
+      <img :src="typeof value == 'string' ? value : imageBase64">
+      <button v-if="typeof value == 'object' || candelete" class="delete is-large" v-on:click="removeFile()" :disabled="disabled || isProcessing">remove</button>
     </figure>
-    <div class="file is-boxed is-small is-centered is-info m-3" v-if="!value">
+
+    <div class="file is-boxed is-centered m-3" v-if="!value && (!value && !imageBase64)" v-bind:class="{'is-info' : !invalid, 'is-danger' : invalid}">
       <label class="file-label">
         <input class="file-input" type="file" ref="file" name="image" :required="required" v-on:input="uploadImage($event.target.value)" accept="image/*" :disabled="disabled || isProcessing">
         <span class="file-cta">
           <span class="file-icon">
             <i class="fas fa-cloud-upload-alt"></i>
           </span>
-          <span class="file-label">
-            Escolha uma imagem...
-          </span>
+          <span class="file-label" v-text="placeholder"></span>
         </span>
       </label>
     </div>
   </div>
 </template>
 <script>
+  
+  const { imageToBase64Async, bytesToSize } = require('../utils.js');
+
   module.exports = {
     props: {
       value: {
-        type: String
+        default: null
+      },
+      placeholder: {
+        type: String,
+        default: "Escolha uma imagem..."
       },
       required: {
+        type: Boolean,
+        default: false
+      },
+      candelete: {
+        type: Boolean,
+        default: true
+      },
+      invalid: {
         type: Boolean,
         default: false
       },
@@ -33,6 +47,11 @@
         type: Boolean,
         default: false
       },
+      maxsize: {
+        type: Number,
+        default: 600 * 1024
+      }
+
     },
     computed: {
       computed_value: function(){
@@ -41,57 +60,77 @@
     },
     data: function() {
       return {
-        isProcessing: false
+        isProcessing: false,
+        imageBase64: null
       };
     },
     methods: {
       uploadImage: function() {
+
         var file = this.$refs.file.files[0];
         if(!file){ return; }
 
-        var formData = new FormData();
-        formData.append("image", file)
+        if(file.size > this.maxsize) {
 
-        this.isProcessing = true;
+          this.$toast.open({
+            type: 'error',
+            message: `O tamanho deste arquivo é ${bytesToSize(file.size)}. O tamanho máximo permitido é ${bytesToSize(this.maxsize)}.`
+          });
 
-        this.$http.post('/rest/media', formData)
-          .then(res => {
-            this.updateValue(res.data);
-          })
-          .catch(res => {
-            console.error(res)
-          })
-          .then(res => {
-            this.isProcessing = false;
-          })
+          return;
+        }
+        
+        this.updateValue(file);
+
+        return;
 
       },
       removeFile: function() {
+
+        if(typeof this.value == 'object') {
+          this.imageBase64 = null;
+          this.updateValue(null);
+          return;
+        }
+
+        var confirm = window.confirm("Tem certeza que deseja excluir este arquivo permanentemente?");
+
+        if(!confirm){ return; }
+
         var file = this.value.split('\/').slice(-1);
 
         this.isProcessing = true;
 
         this.$http.delete(`/rest/media/${file}`)
           .then(res => {
-            this.updateValue(null);
+            this.updateValue(null, true);
           })
-          .catch(res => {
-            if(res.response.status == 404) {
-              this.updateValue(null);
+          .catch((err) => {
+            if(err.response && err.response.status == 404) {
+              this.updateValue(null, true);
             }
           })
           .then(res => {
             this.isProcessing = false;
           })
-
       },
-      updateValue: function (value) {
+      updateValue: function (value, deleteFile = false) {
         let _self = this;
 
-        setTimeout(function(){
-          _self.$emit('input', value ? value : "");
-          _self.$emit('change', value ? value : "");
-        }, 400)
+        if( value ) {
+          imageToBase64Async(value, function(base64){
+            _self.imageBase64 = base64;
+          });
+        }
+
+        _self.$emit('input', value ? value : "");
+        _self.$emit('change', value ? value : "");
+
+        if( deleteFile ) {
+          _self.$emit('deleteupload', "");
+        }
+
+
       }
     },
     created: function() {
